Playerほぼ完成、テスト完了

+++ DragDropAttack/Assets/Script/EnemyData.cs
@@ -8,5 +8,5 @@ public class EnemyData : ScriptableObject
     public string enemyName;    // “G‚Ì–¼‘O
     public int maxHP;           // Å‘åHP
     public int attackPower;     // UŒ‚—Í
-    public int attackTime;      // UŒ‚‚Ü‚Å‚ÌŠÔ
+    public float attackTime;      // UŒ‚‚Ü‚Å‚ÌŠÔ
 }
